--- src/main/java/com/bootproj/pmcweb/Network/Exception/ErrorResponse.java
@@ -1,21 +1,0 @@
-package com.bootproj.pmcweb.Network.Exception;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class ErrorResponse {
-    private String message;
-    private int status;
-    private FieldError errors;
-    private String code;
-
-    @Getter
-    @NoArgsConstructor
-    public static class FieldError{
-        private String field;
-        private String value;
-        private String reason;
-    }
-}
